refactor Scope to a different file
